Add smartcont/auto folder to portable artifacts (#884)
* fix test-validation-session-state test

* make running tests optional in nix builds

* make running tests optional in nix builds

* Revert "fix test-validation-session-state test"

This reverts commit a5bb86918448e234dfcb93cc31c134cd18f123d8.

* add smartcont/auto folder to portable artifacts

* Add check if artifacts were created. If any of tests fail, this is not the case;

+++ assembly/nix/build-linux-arm64-nix.sh
@@ -24,12 +24,14 @@ if [ "$with_tests" = true ]; then
 else
   nix-build linux-arm64-static.nix
 fi
-mkdir artifacts
+
+mkdir -p artifacts/lib
 cp ./result/bin/* artifacts/
+test $? -eq 0 || { echo "No artifacts have been built..."; exit 1; }
 chmod +x artifacts/*
 rm -rf result
 nix-build linux-arm64-tonlib.nix
 cp ./result/lib/libtonlibjson.so.0.5 artifacts/libtonlibjson.so
 cp ./result/lib/libemulator.so artifacts/
-cp -r crypto/fift/lib artifacts/
-cp -r crypto/smartcont artifacts/
+cp ./result/lib/fift/* artifacts/lib/
+cp -r ./result/share/ton/smartcont artifacts/
