+++ assembly/nix/build-macos-nix.sh
@@ -22,12 +22,14 @@ if [ "$with_tests" = true ]; then
 else
   nix-build macos-static.nix
 fi
-mkdir artifacts
+
+mkdir -p artifacts/lib
 cp ./result-bin/bin/* artifacts/
+test $? -eq 0 || { echo "No artifacts have been built..."; exit 1; }
 chmod +x artifacts/*
 rm -rf result-bin
 nix-build macos-tonlib.nix
 cp ./result/lib/libtonlibjson.dylib artifacts/
 cp ./result/lib/libemulator.dylib artifacts/
-cp -r crypto/fift/lib artifacts/
-cp -r crypto/smartcont artifacts/
+cp ./result/lib/fift/* artifacts/lib/
+cp -r ./result/share/ton/smartcont artifacts/
